server 20%
