+++ autocrword/__init__.py
@@ -2,7 +2,6 @@
 
 from . import controllers
 from . import models
-from models import *
 #
 # import sys
 # sys.path.append(r"D:\Program Files (x86)\Odoo 12.0\server")
