+++ autocrword/models/__init__.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
 from . import autoproject, autocivil, autowind, autoeconomic
 from . import autoelectrical
-from . import doc_8
-import numpy as np
-# from chapter_8.MainConstructionQuantitySummarySheet import *
+from addons.Odoo_addons_NB.autocrword.models.source import chapter_5, chapter_8, connect_sql, doc_6
