--- autocrword/models/chapter_5/generate_images.spec
@@ -1,32 +1,0 @@
-# -*- mode: python -*-
-
-block_cipher = None
-
-
-a = Analysis(['generate_images.py'],
-             pathex=['C:\\Users\\Administrator\\PycharmProjects\\Odoo_addons_NB\\autocrword\\models\\chapter_5'],
-             binaries=[],
-             datas=[],
-             hiddenimports=[],
-             hookspath=[],
-             runtime_hooks=[],
-             excludes=[],
-             win_no_prefer_redirects=False,
-             win_private_assemblies=False,
-             cipher=block_cipher,
-             noarchive=False)
-pyz = PYZ(a.pure, a.zipped_data,
-             cipher=block_cipher)
-exe = EXE(pyz,
-          a.scripts,
-          a.binaries,
-          a.zipfiles,
-          a.datas,
-          [],
-          name='generate_images',
-          debug=False,
-          bootloader_ignore_signals=False,
-          strip=False,
-          upx=True,
-          runtime_tmpdir=None,
-          console=False )
