--- autocrword/models/chapter_6/ElectricalCircuit.py
@@ -1,12 +1,0 @@
-class ElectricalCircuit:
-    def __init__(self, project_chapter6_type, single_circuit, double_circuit, buried_cable_35_3, buried_cable_35_1,
-                 tur_number, line_loop_number):
-        self.project_chapter6_type = project_chapter6_type
-        self.single_circuit = single_circuit
-        self.double_circuit = double_circuit
-        self.buried_cable_35_3 = buried_cable_35_3
-        self.buried_cable_35_1 = buried_cable_35_1
-        self.tur_number = tur_number
-        self.line_loop_number = line_loop_number
-
-# 需要消除参数顺序问题
