--- autocrword/models/chapter_8/__init__.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import sys
-
-sys.path.append(r"C:\Users\Administrator\PycharmProjects\Odoo_addons_NB\autocrword\models\chapter_8")
-from models import connect_sql
-from chapter_8 import *
-#
-# __all__ = ['RoundUp']
